Version 2.0 | Menus
Over the last 2 hours, I've successfully built a menuing system for the program. Everything except the program_loop() function has been tested and validated.

The primary purpose of this commit was to store the user's time zones so it wouldn't prompt the user for their time zone EACH and EVERY time. Gosh, it was really getting on my nerves.

Happy too that I was able to clean up the code.

Main looks so good now! (Ｔ▽Ｔ)

The modularization should do well for our cleanliness and future root cause analysis.