Version 2.1 | Settings Validation
Program now checks the integrity of the settings.ini file when settings_menu() is run.

If any of the sections or parameters are missing, the settings.ini file will be replaced and furnished with default values.

Program notifies the user that default values have been installed if this event occurs, encouraging the user to edit the settings.ini file if that is unsatisfactory news in any way.