Version 2.4 | Specific JSON File Retrieval
The current problem I'm running into the program is that it is difficult to retroactively add gym logs. Listen, I got lazy and didn't put every single gym log for the month of July in 2024. In fact, for some reason only one date has been input - which means I've really been slacking. Now that it's August, it's not possible to add a gym log entry to the 07_2024 file. It would be added to the 08_2024 file instead. Is that a horrible thing? I'd say so. The semantic meaning of the logs are that they pertain to gym days that happened in July. If I wanted to reference those gym days, I'd go to the 07_2024 file and be confused that they're not there. I may or may not remember that they were documented retroactively in August and look there for them. Yeah, I could cut and paste the logs from the 08_2024 file once they've been written there - but that's lame and inelegant. That's certainly not a long-term solution.

So, enter new changes. Instead of retrieving the gym log file pertaining to the current month at which the script is being run, the script will instead detect the date of the gym log from the user input and retrieve that corresponding month instead. So even if I am creating the gym log on 08/24/24, the newly created gym log will be written to the 07_2024 file instead.

This is just unilaterally a better design.

Additionally, added more granularity to the logs. Some of the activities I document aren't intense or engaging enough to be considered exercises I would say. For example, before every workout - I do warms up. I wouldn't call that an exercise - or at least put in the same category as the other work-outs. Additionally, before I box I put on my boxing wraps. This isn't an exercise at all. Therefore, I've created the new activity_type variable which identifies the type of activity each entry is. It designates whether they are exercises or not. I've then updated all the semantics of the program and added necessary code to support this change.

Finally, added a couple more QOL features. The start time and end time have now been added to the logs for the viewer's convenience and quality assurance. Also, expanded the exercises classes.

That includes this commit. KBAI. :3