Version 2.5 | Hehehe Forgot to Save :3
Ahoohehehahaeahhoohohohohoehahaheehhohoooooheeeee. : )

Well forgot to mention that there are now two constructs for each "activity log." Exercises and non-exercises. The exercises are the same as they were, moved into their own if statement. The non-exercises are much more brief. Whatever code could be shared between the two constructs were moved out of the if-else chain.

That is all. I hope. >_<